+++ apps/trogon_error/CHANGELOG.md
@@ -2,10 +2,14 @@
 
 ## Unreleased
 
+## v0.2.0
+
+- Add `Trogon.Error.is_trogon_error?/1` guard.
+
 ## v0.1.1
 
-- Fix `t:Trogon.Error.error_opts/0` type
+- Fix `t:Trogon.Error.error_opts/0` type.
 
 ## v0.1.0
 
-- Initial release
+- Initial release.